+++ libnopegl/src/ngpu/vulkan/ctx_vk.h
@@ -55,8 +55,8 @@ struct ngpu_ctx_vk {
     uint32_t cur_image_index;
     int64_t present_time_offset;
 
-    int32_t width;
-    int32_t height;
+    uint32_t width;
+    uint32_t height;
 
     struct darray colors;
     struct darray ms_colors;
@@ -64,7 +64,7 @@ struct ngpu_ctx_vk {
     struct darray rts;
     struct darray rts_load;
     struct ngpu_buffer *capture_buffer;
-    int capture_buffer_size;
+    size_t capture_buffer_size;
     void *mapped_data;
 
     struct ngpu_rendertarget *default_rt;
